python/requirements/prod.txt
flask
Flask-SQLAlchemy
flask-jwt-oidc>=0.1.5
Flask-OpenTracing
jaeger-client